sql: move table metadata update job to new package
This commit creates a new package tablemetadatacache
and moves update_table_metadata_job.go to it. The package
is imported in pkg/upgrade/upgrades to link the job
implementation.

Epic: none

Release note: None
