+++ pkg/sql/tablemetadatacache/BUILD.bazel
@@ -1,0 +1,15 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "tablemetadatacache",
+    srcs = ["update_table_metadata_cache_job.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/tablemetadatacache",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/jobs",
+        "//pkg/jobs/jobspb",
+        "//pkg/settings/cluster",
+        "//pkg/util/log",
+        "@com_github_cockroachdb_errors//:errors",
+    ],
+)
