+++ pkg/upgrade/upgrades/BUILD.bazel
@@ -52,6 +52,7 @@ go_library(
         "//pkg/sql/sem/catconstants",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
+        "//pkg/sql/tablemetadatacache",
         "//pkg/upgrade",
         "//pkg/upgrade/upgradebase",
         "//pkg/util/envutil",
